feat(fs-bridge): add HTTP and Node file system bridges with comprehensive tests
- Implemented HTTPFileSystemBridge with read, exists, listdir operations and error handling.
- Created NodeFileSystemBridge with read, write, exists, listdir, mkdir, and rm operations.
- Developed tests for both bridges covering various scenarios including success and error cases.
- Introduced defineFileSystemBridge utility for creating bridges with options and state management.
- Added TypeScript configuration files for building and type checking.
- Updated playground examples to use the new fs-bridge imports.
- Removed outdated documentation from utils package regarding fs-bridge.
- Updated package.json and tsdown configurations to reflect new structure and dependencies.
