+++ .changeset/tasty-gifts-search.md
@@ -1,0 +1,5 @@
+---
+"@ucdjs/fs-bridge": minor
+---
+
+feat!: migrate fs-bridge from utils to fs-bridge package
