+++ packages/fs-bridge/.gitignore
@@ -1,0 +1,1 @@
+playgrounds/.local
