+++ packages/fs-bridge/eslint.config.js
@@ -1,0 +1,9 @@
+// @ts-check
+import { luxass } from "@luxass/eslint-config";
+
+export default luxass({
+  type: "lib",
+  pnpm: true,
+}, {
+  ignores: ["**/*.md"],
+});
