+++ packages/fs-bridge/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@ucdjs/utils",
-  "version": "0.1.0",
+  "name": "@ucdjs/fs-bridge",
+  "version": "0.0.0",
   "type": "module",
   "author": {
     "name": "Lucas Nørgård",
@@ -13,13 +13,15 @@
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ucdjs/ucd.git",
-    "directory": "packages/utils"
+    "directory": "packages/fs-bridge"
   },
   "bugs": {
     "url": "https://github.com/ucdjs/ucd/issues"
   },
   "exports": {
     ".": "./dist/index.js",
+    "./bridges/http": "./dist/bridges/http.js",
+    "./bridges/node": "./dist/bridges/node.js",
     "./package.json": "./package.json"
   },
   "main": "./dist/index.js",
@@ -39,23 +41,20 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@luxass/unicode-utils-new": "catalog:prod",
     "@luxass/utils": "catalog:prod",
     "@ucdjs/env": "workspace:*",
-    "@ucdjs/fetch": "workspace:*",
+    "@ucdjs/schemas": "workspace:*",
     "defu": "catalog:prod",
-    "memfs": "catalog:prod",
-    "picomatch": "catalog:prod",
     "zod": "catalog:prod"
   },
   "devDependencies": {
     "@luxass/eslint-config": "catalog:linting",
-    "@types/picomatch": "catalog:dev",
     "@ucdjs/tsconfig": "workspace:*",
     "@ucdjs/tsdown-config": "workspace:*",
     "eslint": "catalog:linting",
     "publint": "catalog:dev",
     "tsdown": "catalog:dev",
+    "tsx": "catalog:dev",
     "typescript": "catalog:dev",
     "vitest-testdirs": "catalog:testing"
   },
