+++ packages/fs-bridge/tsconfig.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@ucdjs/tsconfig/base",
+  "include": ["src/**/*", "*.ts", "test"]
+}
