+++ packages/fs-bridge/tsdown.config.ts
@@ -3,6 +3,7 @@ import { createTsdownConfig } from "@ucdjs/tsdown-config";
 export default createTsdownConfig({
   entry: [
     "./src/index.ts",
-    "./src/constants.ts",
+    "./src/bridges/node.ts",
+    "./src/bridges/http.ts",
   ],
 });
