+++ packages/ucd-store/playgrounds/local-playground.ts
@@ -2,7 +2,7 @@
 /* eslint-disable no-console */
 import assert from "node:assert";
 import path from "node:path";
-import NodeFileSystemBridge from "@ucdjs/utils/fs-bridge/node";
+import NodeFileSystemBridge from "@ucdjs/fs-bridge/bridges/node";
 import { createUCDStore, UCDStore } from "../src/store";
 
 const basePath = path.resolve(import.meta.dirname, ".local");
