+++ packages/utils/README.md
@@ -63,67 +63,6 @@ Available filters:
 - `EXCLUDE_README_FILES`: Excludes ReadMe.txt files
 - `EXCLUDE_HTML_FILES`: Excludes all HTML files
 
-### File System Bridge
-
-The filesystem bridge provides a unified interface for file operations across different environments.
-
-#### Node.js File System Bridge
-
-```typescript
-import NodeFileSystemBridge from '@ucdjs/utils/fs-bridge/node';
-
-// Read a file
-const content = await NodeFileSystemBridge.read('/path/to/file.txt');
-
-// Write a file
-await NodeFileSystemBridge.write('/path/to/file.txt', 'Hello World');
-
-// List directory contents
-const files = await NodeFileSystemBridge.listdir('/path/to/dir');
-const allFiles = await NodeFileSystemBridge.listdir('/path/to/dir', true); // recursive
-
-// Create directory
-await NodeFileSystemBridge.mkdir('/path/to/new/dir');
-
-// Check if file exists
-const exists = await NodeFileSystemBridge.exists('/path/to/file.txt');
-
-// Get file stats
-const stats = await NodeFileSystemBridge.stat('/path/to/file.txt');
-console.log(stats.isFile()); // true/false
-console.log(stats.isDirectory()); // true/false
-console.log(stats.size); // file size in bytes
-console.log(stats.mtime); // last modified date
-
-// Remove file/directory
-await NodeFileSystemBridge.rm('/path/to/file.txt');
-await NodeFileSystemBridge.rm('/path/to/dir', { recursive: true, force: true });
-```
-
-#### HTTP File System Bridge
-
-Read-only filesystem bridge for accessing files over HTTP/HTTPS:
-
-```typescript
-import HTTPFileSystemBridge from '@ucdjs/utils/fs-bridge/http';
-
-const httpFS = HTTPFileSystemBridge({
-  baseUrl: 'https://example.com/files/'
-});
-
-// Read remote file
-const content = await httpFS.read('/data/file.txt');
-
-// List remote directory
-const files = await httpFS.listdir('/data/');
-
-// Check if remote file exists
-const exists = await httpFS.exists('/data/file.txt');
-
-// Get remote file stats
-const stats = await httpFS.stat('/data/file.txt');
-```
-
 ## 📄 License
 
 Published under [MIT License](./LICENSE).
