+++ tooling/tsconfig/base.json
@@ -11,9 +11,9 @@
       "@ucdjs/schema-gen": ["./packages/schema-gen/src/index.ts"],
       "@ucdjs/ucd-store": ["./packages/ucd-store/src/index.ts"],
       "@ucdjs/utils": ["./packages/utils/src/index.ts"],
-      "@ucdjs/utils/fs-bridge": ["./packages/utils/src/fs-bridge.ts"],
-      "@ucdjs/utils/fs-bridge/http": ["./packages/utils/src/fs-bridge/http.ts"],
-      "@ucdjs/utils/fs-bridge/node": ["./packages/utils/src/fs-bridge/node.ts"],
+      "@ucdjs/fs-bridge": ["./packages/fs-bridge/src/index.ts"],
+      "@ucdjs/fs-bridge/bridges/http": ["./packages/fs-bridge/src/bridges/http.ts"],
+      "@ucdjs/fs-bridge/bridges/node": ["./packages/fs-bridge/src/bridges/node.ts"],
       "@ucdjs/worker-shared": ["./packages/worker-shared/src/index.ts"],
       "@ucdjs/schemas": ["./packages/schemas/src/index.ts"],
     }
